Adding java doc comments to the new Token providers and related changes (#177)
* Removing an unused method in test.

* A minor fix and some method name changes.

* Minor exception message correction in test.

* Fixing a javadoc comment as it is causing javadoc generation error.

* Adding ADAL4J dependency to release POM. So updating the version to 1.2.1-PREVIEW.

* Adding javadoc comments.

* Fixing a typo in package name.
