+++ azure-servicebus/src/main/java/com/microsoft/azure/servicebus/security/SecurityTokenType.java
@@ -1,7 +1,18 @@
-package com.microsoft.azure.sevicebus.security;
+package com.microsoft.azure.servicebus.security;
 
+/**
+ * Enumeration representing security token types supported by Azure Service Bus.
+ * @since 1.2.0
+ *
+ */
 public enum SecurityTokenType {
+    /**
+     * Shared Access Signature token type
+     */
     SAS(SecurityConstants.SAS_TOKEN_TYPE),
+    /**
+     * JSON web token type
+     */
     JWT(SecurityConstants.JWT_TOKEN_TYPE);
     
     private final String tokenTypeString;
